+++ readthedocs/api/v3/tests/test_notifications.py
@@ -1,6 +1,11 @@
+import django_dynamic_fixture as fixture
+from django.contrib.contenttypes.models import ContentType
 from django.test import override_settings
 from django.urls import reverse
 
+from readthedocs.notifications.constants import CANCELLED, DISMISSED
+from readthedocs.notifications.models import Notification
+from readthedocs.projects.notifications import MESSAGE_PROJECT_SKIP_BUILDS
 from readthedocs.subscriptions.constants import TYPE_CONCURRENT_BUILDS
 from readthedocs.subscriptions.products import RTDProductFeature
 
@@ -23,6 +28,30 @@ class NotificationsEndpointTests(APIEndpointMixin):
         self.assertEqual(response.status_code, 401)
 
         self.client.credentials(HTTP_AUTHORIZATION=f"Token {self.token.key}")
+        response = self.client.get(url)
+        self.assertEqual(response.status_code, 200)
+        self.assertDictEqual(
+            response.json(),
+            self._get_response_dict("notifications-list"),
+        )
+
+        # Adding a CANCELLED/DISMISSED notification won't be returned on this endpoint
+        fixture.get(
+            Notification,
+            attached_to_content_type=ContentType.objects.get_for_model(self.project),
+            attached_to_id=self.project.id,
+            message_id=MESSAGE_PROJECT_SKIP_BUILDS,
+            state=CANCELLED,
+        )
+
+        fixture.get(
+            Notification,
+            attached_to_content_type=ContentType.objects.get_for_model(self.project),
+            attached_to_id=self.project.id,
+            message_id=MESSAGE_PROJECT_SKIP_BUILDS,
+            state=DISMISSED,
+        )
+
         response = self.client.get(url)
         self.assertEqual(response.status_code, 200)
         self.assertDictEqual(